Finishing up Win/Lost Panel.
Added Win and Lost Panel if the condition is met
(Died 10 time, or kill 10 times)
Might add more stuff in the future, but on going testing, if tested is fine, project will temporary stop here.

Gem's remark :
Hello everyone! This is Gem writing Summary. I'm always not that good at these things because I write for my own. This is my first AR project.
I want to point out,
My first experience on Netcode, first experience on AR, and first experience on unity New Input System, so everything is a-bit jangky as I was doing this project for learning, and submission. I have other assignment need to catch on, so I leave this at here first, I might come back and patch thing up in the future, but I'm not sure when, or if I promises it.
Gem's out!